Use Json over Protos for Client-Server protocol.
With the folder structure, there isn't a good place to put protos. Ideally
protos live in a shard folder outside of the server and client but the client
is unable to compile them since they are above the client/src directory. We will
need to switch off of compiling with npm and use something like bazel in order
to have the client use protos. Furthermore, javascript support for protos is
pretty bad and I haven't found good plugins that work with npm start or maven
to generate pre-compiled javascript proto classes.